+++ LabsCM20172/Lab2activities/src/main/java/co/edu/udea/compumovil/gr04_20172/lab2/Add_Apartment_Activity.java
@@ -36,6 +36,11 @@ public class Add_Apartment_Activity extends AppCompatActivity implements View.On
     ByteArrayOutputStream baos;
     byte[] blob;
 
+
+    public Add_Apartment_Activity(){
+        // Required empty public constructor
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
