+++ LabsCM20172/Lab2activities/src/main/res/layout/app_bar_navigation__drawer.xml
@@ -22,14 +22,4 @@
 
     <include layout="@layout/content_navigation__drawer" />
 
-    <android.support.design.widget.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:rippleColor="@color/colorPrimaryDark"
-        android:src="@mipmap/ic_add"
-        />
-
 </android.support.design.widget.CoordinatorLayout>
